requirements.txt
Django==3.2.12
psycopg2-binary>=2.8
django-admin-interface
django-rest-framework
pillow
cryptography
black
isort==5.10.1
pytest==5.4.3
pre-commit
pytest==5.4.3
pytest-black==0.3.10
pytest-cov==2.10.0
pytest-django==3.8.0
pytest-flake8==1.0.6
pytest-freezegun==0.4.1
pytest-isort==2.0.0
pytest-mock==3.5.1
celery==4.3.0
model-bakery==1.12.0
coverage==5.2
flake8==3.8.3
gunicorn==20.0.4
pytz==2020.1
redis==3.5.3
requests==2.26.0
django-celery-beat==1.5.0
django-extensions==3.0.2
jupyter
notebook
drf-access-policy==1.5.0
faker